requirements.txt
Flask==2.2.3
pytube==12.1.2
Gunicorn
mega.py==1.0.8
Werkzeug==2.2.3